+++ src/main/java/denoptim/utils/MoleculeUtils.java
@@ -596,7 +596,7 @@ public class MoleculeUtils
             IAtomContainer mol
     ) {
 
-        ArrayList<Integer> vertIDs = new ArrayList<>();
+        ArrayList<Long> vertIDs = new ArrayList<>();
         for (Vertex v : vertLst) {
             vertIDs.add(v.getVertexId());
         }
@@ -604,7 +604,7 @@ public class MoleculeUtils
         Map<Vertex,ArrayList<Integer>> map = new HashMap<>();
         for (IAtom atm : mol.atoms())
         {
-            int vID = Integer.parseInt(atm.getProperty(
+            long vID = Long.parseLong(atm.getProperty(
                                  DENOPTIMConstants.ATMPROPVERTEXID).toString());
             if (vertIDs.contains(vID))
             {
@@ -618,7 +618,7 @@ public class MoleculeUtils
                 {
                     ArrayList<Integer> atmLst = new ArrayList<>();
                     atmLst.add(atmID);
-                    map.put(v,atmLst);
+                    map.put(v, atmLst);
                 }
             }
         }
@@ -1062,8 +1062,8 @@ public class MoleculeUtils
     {
         IAtomContainer iac = makeSameAs(wholeIAC);
         
-        Set<Integer> wantedVIDs = new HashSet<Integer>();
-        Map<Integer,Vertex> wantedVertexesMap = new HashMap<>();
+        Set<Long> wantedVIDs = new HashSet<Long>();
+        Map<Long,Vertex> wantedVertexesMap = new HashMap<>();
         for (Vertex v : subGraph.getVertexList())
         {
             Object o = v.getProperty(DENOPTIMConstants.STOREDVID);
@@ -1074,8 +1074,8 @@ public class MoleculeUtils
                         + "vertex " + v + ", but is needed to extract "
                                 + "substructure.");
             }
-            wantedVIDs.add(((Integer) o).intValue());
-            wantedVertexesMap.put(((Integer) o).intValue(), v);
+            wantedVIDs.add(((Long) o).longValue());
+            wantedVertexesMap.put(((Long) o).longValue(), v);
         }
         
         // Identify the destiny of each atom: keep, remove, or make AP from it.
@@ -1097,7 +1097,7 @@ public class MoleculeUtils
                         + oriAtm.getSymbol() + wholeIAC.indexOf(oriAtm) 
                         + ", but is needed to extract substructure.");
             }
-            int vid = ((Integer) o).intValue();
+            long vid = ((Long) o).longValue();
             if (wantedVIDs.contains(vid))
             {
                 continue; //keep this atom cpAtm
@@ -1115,7 +1115,7 @@ public class MoleculeUtils
                             + nbr.getSymbol() + wholeIAC.indexOf(nbr) 
                             + ", but is needed to extract substructure.");
                 }
-                int nbrVid = ((Integer) oNbr).intValue();
+                long nbrVid = ((Long) oNbr).longValue();
                 if (wantedVIDs.contains(nbrVid))
                 {
                     // cpAtm is connected to an atom to keep and will therefore
